requisite end-points

+++ app/bank/model.py
@@ -7,8 +7,8 @@ class Requisites(Base):
 
     id = db.Column(db.Integer, primary_key=True)
     bank = db.Column(db.String(100), nullable=False)
-    bik = db.Column(db.Integer, nullable=False, unique=True)
-    rs = db.Column(db.Integer, nullable=False, unique=True)
-    ks = db.Column(db.Integer, nullable=False, unique=True)
+    bik = db.Column(db.BigInteger, nullable=False, unique=True)
+    rs = db.Column(db.BigInteger, nullable=False, unique=True)
+    ks = db.Column(db.BigInteger, nullable=False, unique=True)
     activity = db.Column(db.Boolean, default=False)
     user_id = db.Column(db.Integer, nullable=False)
